requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.20.2
dill==0.3.7
distlib==0.3.7
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.12.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.9.0
future==0.18.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphics.py==5.0.1.post1
greenlet==2.0.2
grpcio==1.54.2
gTTS==2.3.2
gunicorn==21.2.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.12.0
kiwisolver==1.4.4
langchain==0.0.281
langsmith==0.0.33
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.6.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.1
parso==0.8.3
Pillow==9.5.0
pipenv==2023.10.3
platformdirs==3.10.0
playsound==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
PythonTurtle==0.3.2
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.3.3
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
transformers==4.33.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
virtualenv==20.24.5
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.2
zipp==3.16.2
